Refactor 'process_output' entity into its own file
